requirements.txt
asgiref==3.8.1
attrs==24.2.0
channels==4.0.0
channels-redis==4.2.0
daphne==4.0.0
Django>=4.2
django-cors-headers==4.4.0
djangorestframework==3.15.2
psycopg2-binary==2.9.9
redis==5.0.8
chardet==5.2.0
langdetect==1.0.9
